Combine loaders into single module
