+++ sonoradar.html
@@ -200,14 +200,13 @@
     } from './modules/wsManager.js';
 
     import { initZoomControls } from './modules/zoomControl.js';
-    import { initFileLoader, getWavSampleRate } from './modules/fileLoader.js';
+    import { initFileLoader, initDragDropLoader, getWavSampleRate } from './modules/loader.js';
     import { initBrightnessControl } from './modules/brightnessControl.js';
     import { initFrequencyHover } from './modules/frequencyHover.js';
     import { cropWavBlob } from './modules/cropAudio.js';
     import { drawTimeAxis, drawFrequencyGrid } from './modules/axisRenderer.js';
     import { initExportCsv } from './modules/exportCsv.js';
     import { initTrashProgram } from './modules/trashProgram.js';
-    import { initDragDropLoader } from './modules/dragDropLoader.js';
     import { initMapPopup } from './modules/mapPopup.js';
     import { initSidebar } from './modules/sidebar.js';
     import { initTagControl } from './modules/tagControl.js';
